sparse-index: tests for sparse index `reset`
New and updated "sparse checkout compatibility" tests are intended to ensure
all index-affecting modes for reset (all except `reset --soft`) behave
consistently across full and sparse (cone mode and non-cone mode) checkouts.
These tests include verification that the full index is not expanded for any
reset mode in a sparse index checkout.

Additionally, to track behavior for future work, a test is also added to show
that files outside of a sparse checkout definition may be added to the workng
copy of a repository in some cases of a mixed reset.

Signed-off-by: Victoria Dye <vdye@github.com>