t7900: remove use of systemd-analyze
To ensure we have a fully cross-platform test suite, the maintenance
scheduler tests have focused on testing that the way we interact with
platform-specific schedulers is fixed using generic tools such as
'test_cmp' on files. This requires that the developer verifies the
interaction with each platform every time there is a change, but
developers who do not intend to change that interface can be sure they
are not affecting it.

One major way that interaction is protected is given by the
GIT_TEST_MAIN_SCHEDULER environment variable, which redirects the
subprocesses to a script where we can record the selected options.

The systemd interaction implemented in b681b19 (maintenance: add support
for systemd timers on Linux, 2021-09-04) added an extra level of
verification. It used systemd-analyze to verify that the file data is of
the correct format. This was protected by a prerequisite that checked
that systemd-analyze was available on the PATH. However, some versions
of systemd-analyze complain about permissions when called with real
data, so the test fails even though the prerequisite does not.

This change removes the use of systemd-analyze in favor of a forced
comparison with the data written by 'git maintenance start'. Most of the
data is constant, with the only exception being the exact path of the
Git executable to use. To avoid using a relative path from the test
directory, the GIT_EXEC_PATH variable is passed through 'dirname' to get
an absolute path.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>