release: auto-generate release notes
If .github/releases.yml exists, then GitHub will pre-populate the
release notes with a list of the pull requests included. This will
assist in generating the rest of the sections.

If we chose to get fancy with categories, then we could add labels to
our PRs as we merge them. That is likely not important.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>