bundle: safely handle --objects option
Since 'git bundle' uses setup_revisions() to specify the object walk,
some options do not make sense to include during the pack-objects child
process. Further, these options are used for a call to
traverse_commit_list() which would then require a callback which is
currently NULL.

By populating the callback we prevent a segfault in this case. The
command might suffer in performance, but at least it will not have a
hard failure when the user supplies this option. We can also disable
walking trees and blobs in advance of this walk.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>