bundle: alphabetize subcommands better
The usage strings for the 'git bundle' subcommands are not alphabetical.
This also applies to their inspection within cmd_bundle(). Fix this
ordering before we insert a new subcommand.

This change does not reorder the cmd_bundle_*() methods to avoid moving
lines that are more likely wanted in a future 'git blame' call. It is
fine that those longer methods are not ordered alphabetically.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>