remote: move relative_url()
This method was initially written in 63e95beb0 (submodule: port
resolve_relative_url from shell to C, 2016-05-15). As we will need
similar functionality in the bundle URI feature, extract this to be
available in remote.h.

The code is exactly the same. The prototype is different only in
whitespace. The documentation comment only adds explicit instructions on
what happens when supplying two absolute URLs.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>