protocol-caps: implement cap_config()
The 'config' capability sends a list of "key=value" pairs from the
server. These are a set of fixed config values, all prefixed with
"serve." to avoid conflicting with other config values of similar names.

The initial set chosen here are:

* bundle.fetchURI: Allow advertising one or more bundle servers by URI.

* remote.self.promisor: Allow advertising that this server understands
  partial clone.

* remote.self.partialCloneFilter: Advertise one or more recommended
  partial clone filters.

* cone.sparseCheckout[Cone]: Advertise that this repository recommends
  using the sparse-checkout feature [in cone mode].

The client will have the choice to consider these options.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>