commit-graph: document file format v2
The corrected commit date was first documented in 5a3b130ca (doc: add
corrected commit date info, 2021-01-16) and it used an optional chunk to
augment the commit-graph format without modifying the file format
version.

One major benefit to this approach is that corrected commit dates could
be written without causing a backwards compatibility issue with Git
versions that do not understand them. The topological level was still
available in the CDAT chunk as it was before.

However, this causes a different issue: more data needs to be loaded
from disk when parsing commits from the commit-graph. In cases where
there is no significant algorithmic gain from using corrected commit
dates, commit walks take up to 20% longer because of this extra data.

Create a new file format version for the commit-graph format that
differs only in the CDAT chunk: it now stores corrected commit date
offsets. This brings our data back to normal and will demonstrate
performance gains in almost all cases.

Updating the file format version has some downsides. First, Git clients
older than Git 2.22.0 will die() when reading a version it does not
understand. Git clients since then that do not understand this version
will emit a warning about the version, but will fail to parse commits
from the commit-graph file, causing performance degradation.

The commit-graph chunk format allows for flexible ways to add
additional, optional data to the commit-graph file that is silently
ignored by older versions of Git. The BIDX and BDAT chunks are a prime
example, since they include information that accelerates file history
for Git versions that understand them. Older Git versions that do not
understand those chunks are not worse off when reading those chunks.
They are just not as fast as newer Git versions.

With this optional chunk idea in mind, there are a few ways that we
could use new chunks to achieve the data locality of the new generation
number format.

The basic idea in this space is to introduce a new "Commit Data 2" chunk
with ID CDA2.  Since the Commit Data (CDAT) chunk is required by the
commit-graph format, we would need to include it. There are two ways to
make this be compatible with older Git versions:

1. We could keep writing the CDAT chunk, so there are two copies of the
   commit metadata (root tree, commit date, initial parent info) but one
   copy stores topological levels and the other stores corrected commit
   date offsets. Locality during reads are preserved, but this is
   wasteful of disk space.

2. We could write the CDAT chunk as an empty chunk, but this requires
   also writing all zeroes into the OID Fanout (OIDF) chunk and
   creating a new OID Fanout 2 (OIF2) chunk that points into the
   OID Lookup chunk, allowing Git to navigate to the correct rows of
   the CDA2 chunk. This option saves disk space (except for less than
   a kilobyte of wasted data), but causes older Git versions to
   silently ignore the commit-graph file and always parse commits
   from the object database instead.

These two options seem suboptimal for a long-term file format. The
intention of this file format version bump is that it will be disabled
by default. Users who know that they are using only compatible versions
of Git (and third party tools) can opt in to this new version when they
are ready.

Perhaps in the distant future (at minimum a full year since release, but
probably longer) we would want to update the default to use this new
file format version. At that time we would need to gauge the
compatibility of other projects that read commit-graph files as well as
user need.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>