multi-pack-index: normalize --object-dir input
The --object-dir argument to 'git multi-pack-index' allows a user to
specify an alternate to use instead of the local $GITDIR. This is used
by third-party tools like VFS for Git to maintain the pack-files in a
"shared object cache" used by multiple clones.

On Windows, the user can specify a path using a Windows-style file path,
such as "C:\Path\To\ObjectDir". This same path style is used in the
.git/objects/info/alternates file, so it already matches the path of
that alternate.

The tricky part is that we sometimes pass this path as a literal to a
child process, such as 'git pack-objects'. When using forward slashes in
such a path, the 'git multi-pack-index expire' command no longer works.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>