fixup! scalar: do initialize `gvfs.sharedCache`
The scalar shared object cache is a concept borrowed from VFS for Git.
It is used specifically by the git-gvfs-helper and should not be
configured by a "vanilla" scalar clone. Move the logic into a case where
the GVFS Protocol is being used.

I noticed this when I was setting up my new machine and made a "scalar
clone" of git/git. For some reason it was failing to recognize commits
that were already downloaded and was redownloading them.

At minimum, if we were to send an idea of the shared object cache
upstream, it would not be in the way it is currently designed. It would
most likely be in the form of a bare repo where 'scalar clone' creates a
worktree from that bare repo instead of a brand new repo. The biggest
violation that our current design has is that we add objects to and
maintain the files within an alternate.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>