pack-bitmap: add trace2 region around loading bitmaps
The commit lookup table extension in the .bitmap file intends to reduce
the static cost of reading all of the bitmaps to find which correspond
to which commits.

Add a trace2 region so we can measure that cost directly.

This region needs to be closed regardless of the error result, so
refactor all returns to modify a 'result' variable and goto the 'out'
label.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>