pack-bitmap: add timer around loading bitmaps
When using bitmaps for a commit walk, Git looks for a bitmap at each
commit as it walks. If it finds a bitmap, then that bitmap is loaded and
all future commit visits are checked against the bitmap to see if it is
already contained in the bitmap, allowing us to short-circuit the commit
walk. Commits outside the bitmap may also have their own bitmaps which
are then merged with the walk's bitmap using a union operation.

The .bitmap file may contain a commit lookup table extension. By
default, Git will use that extension to help find bitmaps for commits
instead of loading the full list of bitmaps in advance, at least since
28cd730680d (pack-bitmap: prepare to read lookup table extension,
2022-08-14).

However, in a sufficiently large repository and a .bitmap file with 500+
bitmaps, the end-to-end time for 'git rev-list --count
--use-bitmap-index HEAD' slowed down, even though it was saving 11 ms on
loading that initial set of bitmaps.

To investigate, add a new set of timers around loading a bitmap for a
commit. There was previously a trace2 region around
lazy_bitmap_for_commit(), but this region could be emitted thousands of
times during a commit walk, and that causes performance issues all on
its own. By using trace2 timers, we can see how much time is spent in
this region as well as how often we visit this code section.

Add a trace2 counter after lazy_bitmap_for_commit() to track how often
we navigate that code path versus loading a bitmap directly from the
cached hashset.

We can now guarantee that we spend a similar amount of time loading lazy
bitmaps from the lookup table as we do without the lookup table. In the
example 'git rev-list' command in the large repository mentioned
earlier, this counter measured 491 instances of this stopwatch with the
time spent in the method between 115 and 125 ms in both cases. Thus, we
have similar performance in the commit walk.

With these changes, we can replace the previous test that checked for
the trace2 region to use the timer and counter instead.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>