test-lib-functions: add test_timer_count, test_counter
Similar to test_region, these helpers wrap the 'grep' commands that
check that a GIT_TRACE2_EVENT log contains output for a specific trace2
timer or trace2 counter. In each case, we care about the final count for
each of those values. We ignore the timing data from the timer since
that will not be consistent across runs.

These will be used in a test in the next change.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>