scalar: add --no-src option
Some users have strong aversions to Scalar's opinion that the repository
should be in a 'src' directory, even though it creates a clean slate for
placing build outputs in adjacent directories.

The --no-src option allows users to opt-out of the default behavior. Its
documentation indicates that this is a toggle for the default behavior
of creating the 'src' directory, and lists the reasons for that default
so readers can consider those reasons before jumping to change the
default.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>