scalar: diagnose should not retry
When 672196a3073 (scalar-diagnose: use 'git diagnose --mode=all',
2022-08-12) updated 'scalar diagnose' to run 'git diagnose' as a
subprocess, it was passed through the run_git() caller. Upstream, this
caller does not repeat, but in microsoft/git we have a retry loop
specifically for cases where 'git config' commands fail due to
concurrency.

If the underlying 'git diagnose' command fails, then the failure is
repeated. Remove this retry loop for this case.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>