fixup! scalar diagnose: include shared cache info
When this loop was adapted into diagnose.c, it gained an error response
when failing to read a file. This started causing consistent failures
for users with a shared object cache, demonstrating a gap in our testing
support.

The error message itself is confusing:

error: could not read '[...]/info': Permission denied
fatal: unable to create diagnostics archive [...].zip: Permission denied

The "Permission denied" is a red herring for both of these. The second
one is due to a die_errno() even though it's not related to that
filename. The first is actually due to trying to _read the contents of
the directory_.

Since we previously did not error out in this case, we never noticed
this problem in the past. Now, it is causing 'scalar diagnose' to fail.

The fix is to add the filename to the info directory path and read from
that file. Be careful that the entry is actually a file.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>